Fixed issue in implementation of #282 The time formatting function crashed, if used with a date larger than year 2038. This was fixed.
git-svn-id: https://svn.code.sf.net/p/numere/source/trunk@399 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ kernel/core/utils/tools.cpp
@@ -1698,13 +1698,13 @@ bool isToStringArg(const string& sExpr, unsigned int nPos)
 }
 
 // Casts doubles to integers and avoids rounding errors
-int intCast(double number)
+long long int intCast(double number)
 {
 	// if quite close, use rint
 	if (fabs(number - rint(number)) < 1e-7)
 		return rint(number);
 	// otherwise truncate
-	return (int)number;
+	return (long long int)number;
 }
 
 // This function is a wrapper for the usual wcstombs function, which can handle wstrings
